requirements.txt
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
Django==5.1.6
h11==0.14.0
idna==3.10
lxml==5.3.1
numpy==2.2.3
outcome==1.3.0.post0
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
pandas
numpy
matplotlib
scikit-learn
joblib
scipy
djangorestframework
